# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sarah Chatbot API
  description: |
    REST API for the Sarah premium chatbot platform.
    Includes endpoints for widget configuration, chat management, agent dashboard,
    VoIP calls, leads, analytics, AI caller, wallboard, and administration.
  version: 1.0.0
  contact:
    name: Redegal
    url: https://redegal.com
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://sarah.redegal.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Agent JWT token obtained from POST /api/agents/login
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for widget/external integrations
  schemas:
...
    WidgetConfig:
      type: object
      properties:
        branding:
          type: object
          properties:
            companyName:
              type: string
              example: Redegal
            logoUrl:
              type: string
            poweredByText:
              type: string
            showPoweredBy:
              type: boolean
        colors:
          type: object
          properties:
            primary:
              type: string
              example: "#007fff"
            gradientFrom:
              type: string
            gradientTo:
              type: string
        features:
          type: object
          properties:
            enableVoip:
              type: boolean
            enableFileUpload:
              type: boolean
            enableCsat:
              type: boolean
            enableLanguageSelector:
              type: boolean
        layout:
          type: object
          properties:
            position:
              type: string
              enum: [bottom-right, bottom-left]
            width:
              type: integer
            maxHeight:
              type: integer
        i18n:
          type: object
          properties:
            defaultLanguage:
              type: string
            availableLanguages:
              type: array
              items:
                type: string
        businessHours:
          type: object
          properties:
            timezone:
              type: string
            schedule:
              type: array
              items:
                type: object
...
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, converted, lost]
        score:
          type: integer
        businessLine:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, agent, supervisor]
        email:
          type: string
        languages:
          type: array
          items:
            type: string
        businessLines:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [online, busy, away, offline]
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 100
        password:
          type: string
          minLength: 4
          maxLength: 200
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        agent:
          $ref: '#/components/schemas/Agent'
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        secret:
          type: string
        created_at:
          type: string
          format: date-time
    CannedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shortcut:
          type: string
          example: /hello
        title:
          type: string
        content:
          type: string
        language:
          type: string
        business_line:
          type: string
        category:
          type: string
        usage_count:
          type: integer
    Theme:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        config:
          $ref: '#/components/schemas/WidgetConfig'
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/config/widget:
    get:
      tags: [Config]
      summary: Get widget theme configuration
      description: Public endpoint called by the widget loader to fetch theme, branding, and feature configuration.
      parameters:
        - name: tenant
          in: query
          schema:
            type: string
          description: Tenant slug (defaults to "redegal")
      responses:
        '200':
          description: Widget configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetConfig'
...
  /api/config/theme:
    get:
      tags: [Config]
      summary: List all themes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of themes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Theme'
  /api/config/theme/{id}:
    put:
      tags: [Config]
      summary: Update a theme
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config:
                  $ref: '#/components/schemas/WidgetConfig'
      responses:
        '200':
          description: Updated theme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theme'
        '404':
          description: Theme not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/config/tenants:
    post:
      tags: [Config]
      summary: Create a new tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                name:
                  type: string
                domain:
                  type: string
      responses:
        '200':
          description: Created tenant
          content:
            application/json:
              schema:
                type: object
  /api/config/canned:
    get:
      tags: [Config]
      summary: List canned responses
      security:
        - BearerAuth: []
      parameters:
        - name: language
          in: query
          schema:
            type: string
        - name: businessLine
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Array of canned responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CannedResponse'
    post:
      tags: [Config]
      summary: Create a canned response
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shortcut, title, content]
              properties:
                shortcut:
                  type: string
                title:
                  type: string
                content:
                  type: string
                language:
                  type: string
                businessLine:
                  type: string
                category:
                  type: string
      responses:
        '200':
          description: Created canned response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CannedResponse'
  /api/config/canned/{id}:
    delete:
      tags: [Config]
      summary: Delete a canned response
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/config/webhooks:
    get:
      tags: [Config]
      summary: List webhooks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
    post:
      tags: [Config]
      summary: Create a webhook
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                secret:
                  type: string
      responses:
        '200':
          description: Created webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'

  # ═══════════════════════════════════════════════════════
  # CHAT
  # ═══════════════════════════════════════════════════════
...
  /api/leads:
    get:
      tags: [Leads]
      summary: List leads
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, converted, lost]
        - name: businessLine
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Array of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
  /api/leads/{id}:
    patch:
      tags: [Leads]
      summary: Update a lead
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, contacted, qualified, converted, lost]
                notes:
                  type: string
                  maxLength: 5000
      responses:
        '200':
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/leads/{id}/score:
    post:
      tags: [Leads]
      summary: Score a lead
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lead score
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: integer

  # ═══════════════════════════════════════════════════════
  # AGENTS
  # ═══════════════════════════════════════════════════════
  /api/agents/login:
    post:
      tags: [Agents]
      summary: Agent login
      description: Authenticates an agent and returns a JWT token. Rate limited to 5 attempts per minute per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agents/me:
    get:
      tags: [Agents]
      summary: Get current agent profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
  /api/agents/status:
    patch:
      tags: [Agents]
      summary: Update agent status
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [online, busy, away, offline]
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/agents:
    get:
      tags: [Agents]
      summary: List online agents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of agents (without password hashes)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
  /api/agents/queue:
    get:
      tags: [Agents]
      summary: Get waiting conversations queue
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of waiting conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ═══════════════════════════════════════════════════════
  # CALLS
  # ═══════════════════════════════════════════════════════
...